rename alu_instructions into alu_operations
